Chore/Add elastic-apm to dependencies
It seemed to be missing, and Django won't start without it because it's in
INSTALLED_APPS

NOTE: I suspect pyproject.toml and poetry.lock files have become out of
sync with the requirements.txt files, e.g. Django is 3.2.25 in poetry.lock, but
4.2.15 in requirements/prod.txt.

It might be a small project to link everything back up. For now, have just
made these changes manually in the requirements files, which is how I suspect
other recent changes have been made.